considering there was a re-rendering error that causes the weather app to vanish, a pages folder was created to create a ToDoPage where all the data assoicated with to-do app was moved from the app to that page file to isolate the state management thus prevent state changes and re-renders to the weather app
